Refactor the way we check the divs while running
Previously in order to update the correct div when updating the corresponding cell we would make two queries, first selecting all divs with a data-x that is equal to the cell's position i and then looking through that node for a div with a data-y == x. That was both wasteful and confusing to read.

Instead, now we give each div and ID on creation and can access it through it, cutting down on computation time. The ID corresponds to the cell's i and x co-ordinate, so the cell in board[3][13] is connected to the div with an ID of 313.