Add random button, remove clear, bug fix, refactor
Random button is added to replace the clear button. The random button generates a random state for the grid using the given parameters for grid size.

Clear button was removed as it's function is identical to the Create Grid button.

Fixed a bug where clicking Create Grid or Clear(now Random) would not reset the counter for the generations, and instead would increment it.

The stop helper function was removed as it was not used anywhere. Added resetCounters function to handle resetting the counters where appropriate.