Fix id bug
With the way ids for divs where defined some of them ended up with the same id. For example the div at position x 1 and y 14 had an id of 114, but the div at x 11 y 4 ALSO had an id of 114. This caused cells to jump around the board as they would look at the first div with the required id to connect to. I added a - in the id declaration so that x1 y14 would now be 1-14 and x11 y4 will be 11-4. This ensures ids are unique and the game works as expected now.